pyproject.toml
[tool.poetry]
name = "microservices-platform"
version = "1.0.0"
description = "Production-ready e-commerce microservices platform"
authors = ["Team <team@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# Web Framework
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"

# Database - PostgreSQL
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
alembic = "^1.12.0"

# Database - MongoDB
motor = "^3.3.0"
pymongo = "^4.5.0"

# Cache & Session
redis = "^5.0.0"
hiredis = "^2.2.0"

# Message Brokers
aiokafka = "^0.8.1"
aio-pika = "^9.3.0"

# Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

# HTTP Client
httpx = "^0.25.0"
aiohttp = "^3.9.0"

# Utilities
python-dotenv = "^1.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.10.0"
ruff = "^0.1.3"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "services"]
